refactor: centralize openai usage
