+++ src/lib/openai.ts
@@ -5,5 +5,6 @@ if (!process.env.OPENAI_API_KEY) {
 }
 
 export const openai = new OpenAI({
-  apiKey: process.env.OPENAI_API_KEY,
-}); 
+  apiKey: process.env.OPENAI_API_KEY!,
+  organization: process.env.OPENAI_ORGANIZATION || undefined,
+});
